requirement.txt
Django==1.11
argparse==1.2.1
djangorestframework==3.6.2
ipython==5.3.0
pathlib2==2.2.1
pickleshare==0.7.4
pytz==2017.2
traitlets==4.3.2
wsgiref==0.1.2
